requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-gis==0.17
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
